[Feat] 인증 관련 API auth 디렉토리로 옮김
